observability/scrapeAoAIPricing/requirements.txt
requests
beautifulsoup4
azure-ai-inference
python-dotenv
azure-identity
azure-storage-blob